third_party/blink/web_tests/external/wpt/html/canvas/element/text/2d.text.direction.valid-expected.txt

This is a testharness.js-based test.
[FAIL] Canvas test: 2d.text.direction.valid
  assert_equals: ctx.direction === 'inherit' (got ltr[string], expected inherit[string]) expected "inherit" but got "ltr"
Harness: the test ran to completion.
